backend/pyproject.toml
[project]
name = "visa-kit-backend"
version = "0.1.0"
description = "Spain digital nomad visa application kit — backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    # ORM and validation
    "sqlmodel>=0.0.22",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    # Document templates
    "docxtpl>=0.18",
    "python-docx>=1.1",
    "pypdf>=5.1",
    "reportlab>=4.2",
    # External services
    "anthropic>=0.40",
    "httpx>=0.27",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # Utilities
    "python-multipart>=0.0.12",
    "email-validator>=2.2",
]

[project.optional-dependencies]
prod = [
    "psycopg2-binary>=2.9",
    "boto3>=1.35",
]

dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "black>=24",
    "ruff>=0.7",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
